hyhull now uses the hydra-role-management gem for providing a simple ui for role management.  The Role model has been extended to allow the classification of RoleTypes:  - hyhull (For application type roles, like contentCreator etc..)  - user (Specify if user is staff, student or other...)  - department (Store the department of a user)  - faculty (Store the faculty of a user)
This will enable easier management of roles, and the ability to add resource permissions based on the users different roles.

A new Person model has been added to model the person_db_table that hyhull will use for downloading accounts.  In dev and test mode, this is just a standard
db table, in production it will be a view.

Still to add before merging into master:-
 - Keys on Roles DB to ensure that multiple copies cannot be inserted into the table
 - RSpec coverage for new functionality in User, Person, Role, RoleType models
 - Modify the Role-management views to filter 'non-editable' roles_users (i.e. staff/student etc..), but allow the adding/modifying of hyhull roles